implement make a triangle mesh sprain system and update demo2 to create and display a triangle mesh sprain system

+++ src/demos/demo2.py
@@ -28,7 +28,8 @@ def ui_callback(state: dict):
     if state['ui_is_running']:
         _update_system(state)
         _update_polyscope_mesh(state)
-        ps.screenshot(transparent_bg=True)
+        if (state['record']):
+            ps.screenshot(transparent_bg=False)
 
 
 def ui_system_parameters(state: dict):
@@ -90,7 +91,7 @@ def _initialize_system(state: dict):
     """
     Initialize the system with specified point mass and spring stiffness
     """
-    state['system'] = demo_systems.make_triangle_mesh_system(
+    state['system'] = demo_systems.make_triangle_mesh_strain_system(
         state['mesh'], state['ui_point_mass'], state['ui_spring_stiffness'])
 
 
@@ -130,19 +131,20 @@ def _initialize_polyscope_mesh(state: dict):
     Initialize the polyscope mesh to be a curve network where each edge corresponds to a spring in our system.
     """
     system = state['system']
+    mesh = state['mesh']
     edges = np.array([edge(c) for c in system.cons])
-    state['ps_mesh'] = ps.register_curve_network(
-        name='ps_mesh', nodes=system.q, edges=edges)
+    state['ps_mesh'] = ps.register_surface_mesh(
+        name='ps_mesh', vertices=mesh.points(), faces=mesh.face_vertex_indices())
 
 
 def _update_polyscope_mesh(state: dict):
     """
     Update the polyscope curve network according to the current vertex positions of the system.
     """
-    state['ps_mesh'].update_node_positions(state['system'].q)
+    state['ps_mesh'].update_vertex_positions(state['system'].q)
 
 
-def main(filename: str):
+def main(args: dict):
     """
     Creates a Polyscope application visualizing a mass-spring system simulated using projective dynamics.
 
@@ -151,7 +153,7 @@ def main(filename: str):
     """
 
     # Read the mesh
-    mesh = om.read_trimesh(filename)
+    mesh = om.read_trimesh(args.filename)
 
     # Set the application name
     ps.set_program_name('Projective Dynamics Demo')
@@ -174,7 +176,8 @@ def main(filename: str):
         'ps_mesh': None,
         'ui_is_running': False,
         'ui_h': 0.01,
-        'ui_steps_per_frame': 10
+        'ui_steps_per_frame': 10,
+        'record' : args.record
     }
     # Initialize the system and polysope mesh
     _initialize_system(state)
@@ -199,12 +202,20 @@ if __name__ == '__main__':
     parser.add_argument("filename",
                         help="the input mesh filename",
                         type=str)
+    
+    # Optional argument to screen 
+    parser.add_argument("--record", 
+                        help="whether or not to record the application screen",
+                        action="store_true")
 
     # Parse the command line arguments
     args = parser.parse_args()
-
-    # Pass the filename to our main function
-    main(args.filename)
-    subprocess.run(
-        f"mkdir -p ../demos/ &&ffmpeg -y -framerate 30 -pattern_type glob -i 'screenshot_*.png' -c:v libx264 -pix_fmt yuv420p ../demos/{os.path.basename(args.filename)}.mp4 && rm screenshot_*.png",
-        shell=True)
+    
+    # Pass the args to our main function
+    main(args)
+    
+    # Optionally record the application screen
+    if (args.record):
+        subprocess.run(
+            f"mkdir -p ../demos/ &&ffmpeg -y -framerate 30 -pattern_type glob -i 'screenshot_*.png' -c:v libx264 -pix_fmt yuv420p ../demos/{os.path.basename(args.filename)}.mp4 && rm screenshot_*.png",
+            shell=True)
